% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\name{Mega2ENVGenABEL}
\alias{Mega2ENVGenABEL}
\title{generate gwaa.data-class object}
\usage{
Mega2ENVGenABEL(markers = NULL, force = TRUE, makemap = FALSE,
  sort = TRUE, mapno = 0, envir = ENV)
}
\arguments{
\item{markers}{data frame of markers being processed}

\item{force}{pass value to gwaa}

\item{makemap}{pass value to gwaa}

\item{sort}{pass value to gwaa}

\item{mapno}{specify which map index to use for genetic distances}

\item{envir}{"environment" containing SQLite database and other globals}
}
\value{
gwaa.class-object of previously read.Mega2DB database
}
\description{
Directly write a gwaa.data-class object from ENV tables
}
\examples{
\dontrun{
ENV <- read.Mega2DB("my.db")

gwaa = Mega2ENVGenABEL(NULL)
str(gwaa)
head(summary(gwaa))
}
}
