% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\docType{package}
\name{Mega2GenABEL-package}
\alias{Mega2GenABEL-package}
\title{Mega2GenABEL package}
\description{
This package processes R data frames that represent a Mega2R SQLite3 database and
generates the corresponding GenABEL gwaa.data-class object.
}
\author{
Robert V Baron
}
