% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\name{Mega2GenABEL}
\alias{Mega2GenABEL}
\title{generate required VCF header}
\usage{
Mega2GenABEL(prefix, markers = NULL, mapno = 0, envir = ENV)
}
\arguments{
\item{prefix}{prefix for vcf file name}

\item{markers}{data.frame of markers being processed}

\item{mapno}{specify which map index to use for genetic distances}

\item{envir}{"environment" containing SQLite database and other globals}
}
\value{
gwaa.class-object of previously read(.Mega2DB) database
}
\description{
Call functions to: create .tped file, .tfam file and .phe file.
 Call the GenABEL functions to process these files.
}
\examples{
\dontrun{
ENV <- read.Mega2DB("my.db")

Mega2GenABEL(prefix, NULL)
}
}
