% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\name{Mega2GenABELcoding}
\alias{Mega2GenABELcoding}
\title{generate GenABEL coding vector}
\usage{
Mega2GenABELcoding(markers = NULL, Freq.x, envir = ENV)
}
\arguments{
\item{markers}{data frame of markers being processed}

\item{Freq.x}{frequency of first allele calculated from actual data.}

\item{envir}{"environment" containing SQLite database and other globals}
}
\value{
None
}
\description{
Each element of the coding vector is a lookup in the GenABEL:::alleleID.codes() array.
 The alleles are ordered so that with the greater frequency appears first.
}
\examples{
\dontrun{
Mega2GenABELcoding(envir)
}
}
