% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\name{Mega2GenABELconvert}
\alias{Mega2GenABELconvert}
\title{generate GenABEL compressed genotype matrix}
\usage{
Mega2GenABELconvert(markers = NULL, envir = ENV)
}
\arguments{
\item{envir}{"environment" containing SQLite database and other globals}
}
\value{
None
}
\description{
The matrix is (# of samples / 4 ) x (# of markers).  (Round samples to multiple of 4.
 Each byte stores data for 4 samples; intfn() generates the 4 - 2 bit encodings.
}
\examples{
\dontrun{
Mega2GenABELconvert(envir)
}
}
\keyword{internal}
