% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2tst.R
\name{tstph}
\alias{tstph}
\title{test phenotype fields}
\usage{
tstph(mega_ = mega, gwaa_ = srdta)
}
\arguments{
\item{mega_}{name of first gwaa.data-class object}

\item{gwaa_}{name of first gwaa.data-class object}
}
\value{
None
}
\description{
verify that the fields in two gwaa.data-class objects have the same phenotype fields
}
\examples{
\dontrun{
tstph()
}
}
