% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2tst.R
\name{dmpPed}
\alias{dmpPed}
\title{generate required .ped, .fam and .map  for PLINK PED files}
\usage{
dmpPed(gwaa_ = srdta, pfx = "srdta", default = "bt")
}
\arguments{
\item{gwaa_}{name of gwaa.data-class object used as input}

\item{pfx}{prefix for PLINK .ped/.map/.phe file names}

\item{default}{name of phenotype used for the 6th column of .ped file}
}
\value{
None
}
\description{
Use provided gwaa.class-object and create a PLINK .ped file, PLINK .map file
 and a PLINK .phe (phenotypes) file.  By default, srdta (a sample from GenABEL)
 is used for the gwaa.class-object.  The files are generated with a prefix of
 "srdta" unless a pfx argument is provided.
 NOTE: These PLINK files may be used by the Mega2 executable to produce a database.
}
\examples{
\dontrun{
dmpPed()
or
dmpPed(mygwaa, "name", "cc")
}
}
