% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2tst.R
\name{dmpPed}
\alias{dmpPed}
\title{generate required .ped, .fam and .map  for PLINK PED files}
\usage{
dmpPed(gwaa_ = srdta, pfx = "srdta")
}
\arguments{
\item{gwaa_}{name of gwaa.data-class object}

\item{pfx}{prefix for PLINK ped file names}
}
\value{
None
}
\description{
Use provided gwaa.class-object and create a PLINK PED file, PLINK MAP file
 and a PLINK phe file to hold the phenotypes.
}
\examples{
\dontrun{
dmpPed()
}
}
