% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\docType{package}
\name{Mega2GenABEL-package}
\alias{Mega2GenABEL-package}
\title{Mega2GenABEL package}
\description{
This package reads a Mega2 SQLite3 database into R dataframes and
generates input data for GenABEL from these same frames.
}
\author{
Robert V Baron
}
