% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\name{mkGenABELphe}
\alias{mkGenABELphe}
\title{generate required PLINK (.phe) file}
\usage{
mkGenABELphe(prefix, ENV, markers)
}
\arguments{
\item{prefix}{prefix for vcf file name}

\item{ENV}{"environment" containing SQLite database and other globals}

\item{markers}{data.frame of markers being processed}
}
\value{
None
}
\description{
Generate the .phe (phenotype) file for PLINK which is used by GenAbel.  The person
 must match that specified in the .tfam file
}
\examples{
\dontrun{
mkGenABELphe(prefix, ENV, NULL)
}
}
