% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\name{mkGenABELphe}
\alias{mkGenABELphe}
\title{generate required PLINK (.phe) file}
\usage{
mkGenABELphe(envir)
}
\arguments{
\item{envir}{"environment" containing SQLite database and other globals}
}
\value{
None
}
\description{
Generate the .phe (phenotype) file for PLINK which is used by GenAbel.  The person
 must match that specified in the .tfam file
}
\examples{
\dontrun{
mkGenABELphe(envir)
}
}
