% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\name{mkGenABELtfam}
\alias{mkGenABELtfam}
\title{generate required fam family for PLINK TPED (.tfam) file}
\usage{
mkGenABELtfam(prefix, ENV, markers)
}
\arguments{
\item{prefix}{prefix for vcf file name}

\item{ENV}{"environment" containing SQLite database and other globals}

\item{markers}{data.frame of markers being processed}
}
\value{
None
}
\description{
Generate the six column .tfam file used with the .tped file.  Note: Only the person id
 column appears to be used by GenABEL.
}
\examples{
\dontrun{
mkGenABELtfam(prefix, ENV, NULL)
}
}
