% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\name{mkGenABELtfam}
\alias{mkGenABELtfam}
\title{generate required fam family for PLINK TPED (.tfam) file}
\usage{
mkGenABELtfam(prefix, envir)
}
\arguments{
\item{prefix}{prefix for generated file name}

\item{envir}{"environment" containing SQLite database and other globals}
}
\value{
None
}
\description{
Generate the six column .tfam file used with the .tped file.  Note: Only the person id
 column appears to be used by GenABEL.
}
\examples{
\dontrun{
mkGenABELtfam(prefix, envir)
}
}
