% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega2genabel.R
\name{mkGenABELtped}
\alias{mkGenABELtped}
\title{generate a PLINK TPED file for GenABEL}
\usage{
mkGenABELtped(prefix, ENV, markers = NULL, mapno = 0)
}
\arguments{
\item{prefix}{prefix for .tped file name}

\item{ENV}{"environment" containing SQLite database and other globals}

\item{markers}{markers selected to be in output file}

\item{mapno}{specify which map index to use for genetic distances}
}
\value{
None
}
\description{
Generate a PLINK TPED file from the specified Mega2 SQLite database.  The file is named "prefix".tped
 If the markers arg is.null(), the entire ENV$markers set is used otherwise markers arg MUST
 be a subset of the ENV$markers data.frame -- same columns, but pruned rows.
}
\examples{
\dontrun{
mkGenABELtped("foo")

mkGenABELtped("foo", ENV$markers[ENV$markers$chromosome >= 20,])
}
}
